+++ CHANGELOG.md
@@ -2,6 +2,8 @@ pnet-data-api 2.0.3
 ====================
 
 * Add autoComplete to applications.
+* Add autoComplete to activities.
+* Add autoComplete to functions.
 
 pnet-data-api 2.0.2
 ===================
@@ -28,6 +30,7 @@ pnet-data-api 1.21.3
 
 * Add autoComplete to applications.
 * Add autoComplete to activities.
+* Add autoComplete to functions.
 
 pnet-data-api 1.21.2
 ====================